+++ backend/src/main/kotlin/com/swpp/footprinter/domain/trace/dto/InitialTraceDTO.kt
@@ -1,7 +1,7 @@
 package com.swpp.footprinter.domain.trace.dto
 
-import com.swpp.footprinter.domain.photo.model.PhotoInitialTraceDTO
-import com.swpp.footprinter.domain.place.model.PlaceInitialTraceDTO
+import com.swpp.footprinter.domain.photo.dto.PhotoInitialTraceDTO
+import com.swpp.footprinter.domain.place.dto.PlaceInitialTraceDTO
 import java.util.*
 
 data class InitialTraceDTO(
